Initial Commit for EAG for GoLang
EAG stands for Endpoint Activity Generator and is used to create
activity on some endpoint. This commit contains the shell and file
operations.

The shell is one way (the only way at the moment) that the user
will run any of the activities.

File operations include: Create, Delete, Write-Append,
Write-Overwrite, and Read.